refactor: rename zeebe BackupStoreCfg to BackupCfg
