+++ zeebe/qa/integration-tests/src/test/java/io/camunda/zeebe/it/cluster/backup/AzureRestoreAcceptanceIT.java
@@ -10,7 +10,7 @@ package io.camunda.zeebe.it.cluster.backup;
 import io.camunda.configuration.Camunda;
 import io.camunda.zeebe.broker.system.configuration.BrokerCfg;
 import io.camunda.zeebe.broker.system.configuration.backup.AzureBackupStoreConfig;
-import io.camunda.zeebe.broker.system.configuration.backup.BackupStoreCfg.BackupStoreType;
+import io.camunda.zeebe.broker.system.configuration.backup.BackupCfg.BackupStoreType;
 import io.camunda.zeebe.test.testcontainers.AzuriteContainer;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.testcontainers.junit.jupiter.Container;
