+++ zeebe/qa/integration-tests/src/test/java/io/camunda/zeebe/it/cluster/backup/S3RestoreAcceptanceIT.java
@@ -11,7 +11,7 @@ import io.camunda.configuration.Camunda;
 import io.camunda.zeebe.backup.s3.S3BackupConfig.Builder;
 import io.camunda.zeebe.backup.s3.S3BackupStore;
 import io.camunda.zeebe.broker.system.configuration.BrokerCfg;
-import io.camunda.zeebe.broker.system.configuration.backup.BackupStoreCfg.BackupStoreType;
+import io.camunda.zeebe.broker.system.configuration.backup.BackupCfg.BackupStoreType;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.test.testcontainers.MinioContainer;
 import java.time.Duration;
